Show "needs build" panel when open the file if applicable
We were stuck in the "Initializing..." panel when reopening a file after
triggering a gradle clean. We need to check whether a build is required when
opening the file in order to display the appropriate panel in that case,
which allows users to trigger a build.

Fixes: 187705125
Test: Manual by closing/reopening a Compose file after gradle clean
Change-Id: Icf3d4fc348ce26201a23fc70fdaaa51177f63fd8
(cherry picked from commit 963011ee51b0833234c3bbcf790142d8f8ca799f)