[Dsl] Handle Int/String Sdk properties
Historically fooSdk properties were all set by an overloaded method
fooSdk(), handling the fact that the argument could be an integer (for a
released Sdk version) or a string (for a preview).  Now that these
properties are split into two, we need some way to model the property,
and in particular to do the right thing if client code wishes to set the
Sdk version from one kind to the other (e.g. in a project with declared
minSdk 30, to change to using a preview for "T" one can no longer simply
change the argument to the application statement).

PropertyTransforms allow making non-local changes to the Dsl tree in
response to changing the values of Dsl properties.  Define a new
transform, SdkOrPreviewTransform, which examines the type of the value
being set and determines which of the Integer, String or generic method
to use, optionally also constrained by version considerations.

Unfortunately, this doesn't quite work, in that after carefully
constructing a Dsl element with the right form, model effect and name,
the name conversion mechanism kicks in at a lower level and rewrites the
element's name to something wrong: the Dsl parser/writer do not have
enough information to be able to determine the correct syntactic form
for this semantic operation.  This might be indicative that there's a
layering violation going on here, but it doesn't seem obvious to me that
there is: it might just be that we need to be able to indicate to the
lower level that we have already taken into consideration issues around
syntax.

In any case, the form we have is unique: a Dsl element with a
ModelEffectDescription acting on a property, whose name is not the same
as the GradleElementName of the underlying element, and whose effect is
CREATE_WITH_VALUE (which is never generated when parsed).  Handle this
case specially when generating an external name.

Bug: 148657110
Bug: 184935026
Test: existing, more in forthcoming commits
Change-Id: I40b4f552517d46bac57f7702ae126bf723ded4fb