Eliminate outputKind callback in GradleApkProvider
by replacing it with a direct call to `useSelectApksFromBundleBuilder`.

Sunce the run configuration is not expected to change its properties
between run is requested and deploy started it is safe to remeber
`alwaysDeployApkFromBundle` in the apk provider and thus hold all
the required values to make a call to `useSelectApksFromBundleBuilder`.

This is a step to remove `outputs` property from `IdeAndroidArtifact`.

Bug: 37797956
Test: ApkProviderIntegrationTest
Test: ApkProviderOldAgpIntegrationTest
Change-Id: I67901a4fc894f9553cbf46ddbb76a0438bbf7495