Do not auto zoom-to-fit in custom view preview and Layout Editor
This CL remove the auto zoom-to-fit function when device or render api
level is changed.

This affects custom view preview and layout editor:

- For custom view preview, it does not auto zoom-to-fit when toggling
  wrap content buttons.

- For layout editor, it does not zoom-to-fit when changing device and
  api level.

Note:
If a file is opened at first time, it will still zoom-to-fit
automatically. Otherwise the last zoom level is applied.

This CL could be considered as a revert of change
Ia4d3e933ddaaf65ab31077b8e8c25aa553919fa6

Fixes: 189081609
Bug: 142254879
Bug: 184391738
Test: manually (UI-related)
Change-Id: I61d888e7a1e21608202e271b3b14ee938078aa5f