Do not register duplicated ResourceChangeListener in SceneManager
It is allowed to activate the SceneManager even it is activated.
When it happens, duplicated listener is registered to
ResourceNotifcationManager.
It may cause a memory leakage issue if the number of deactivation is
less than activation.

After this CL, the activation state depends on activate() and
deactivate() which is called last.

Bug: 188825161
Test: Add regression test.
Change-Id: I069de1fe0135b1489cfbceafb51570236310fa92
(cherry picked from commit 6e59e3e03276ec28a190a8b7827fd660e45f9a6a)