[Dsl] Support pluginManagement in Gradle settings files
PluginManagement has a `plugins { }` block, without all the additional
complexity in a build file with multiple ways of applying plugins.  This
makes the implementation of support for plugins more straightforward,
though because we re-use the existing parsing for simplicity we will
parse some invalid declarations.  This is not a substantial problem, as
we only emit valid plugin declarations.

Bug: 185214653
Test: added
Change-Id: I35c9d3bc88c4d00076c10f930e0109ee1ce9049e