Update root bounds calculation
Previously, when connecting with system node filtering enabled, system
nodes were ignored for the purpose of zoom-to-fit calculations, even if
they were still responsible for painting. This meant that if there were
no non-system nodes occupying the whole width and height of the screen
the view could be zoomed in more than expected.

Fixes: 188770540
Test: added a test
Change-Id: I611f017cacdcfc7d988d009a229fad7ef4735acb