Add more debug logs to Compose Preview refresh listeners
Some users are facing an issue where Compose Previews stop being updated at
some point. From the latest logs provided in issue 189559350, it seems the
refresh() method is not called after a feel refreshes, giving the impression
that both the BuildListener and the ChangeListener we set up in
ComposePreviewRepresentation#onInit are not listen anymore, otherwise they'd
call refresh through buildSucceeded and onDocumentUpdated callbacks
respectively.

This CL adds some logs to verify these callbacks are indeed called, so we
can determine if the problem is that these methods are not being called or
if their call to refresh() is not working for some other reason.

Bug: 189559350
Test: N/A
Change-Id: Icde1415bfc50a8cd3b06e90717c7bfc7cb36ee1e
(cherry picked from commit 49b56f3f9a81fee6cc3c5f779df14b3aa5e58437)