'Build project automatically' checkbox may be both on CompilerConfigurable and on GradleCompilerSettingsConfigurable
IJP-630 (follow-up)
In Android projects, the 'Build project automatically' checkbox is on the GradleCompilerSettingsConfigurable page. In other projects it is
on the CompilerConfigurable page. This fix ensures that the correct 'Build project' option appears on the 'Actions on Save' page.

(cherry picked from commit 211be8dceb825ea5956405ad549d1da347feab8e)

IJ-CR-10224

GitOrigin-RevId: a15de96a4b1d02c85b24243a156e2a732d0ed569