APK Sign: Generate service name properly for saved passwords
One of the root causes found for passwords not being saved properly
is the fact that only one username:password pair is allowed
per "service name" by IJ PasswordSafe password storage.
In KeystoreStep we used PasswordSafe in an old deprecated way
providing requestor class which was used as a service name internally.
As this service name was static each new password saved in this dialog
was overwriting the previous ones globally across all projects.

Fix: change deprecated way of creating CredentialAttributes to a new
recommended way and include password key (which specifies keystore and
key alias) in service name of CredentialAttributes. This way we are able
to safe passwords for all keystorage/key pairs, not just the last one.
Add migration logic to load password from the old place and clear up
the old saved password on save.

Bug: 192344567
Test: Added test cases to KeystoreStepTest.kt
Change-Id: I2101a925aada8a2a8233c227ac94fe9f96d2e340