Use exec configuration for _stamper executable
The 'host' configuration does not support proper toolchain
resolution. This means that the executable always ran in either python2
or python3 depending on the host default. The host default is becoming
python3, and we need the executable to still resolve to python2 based on
the python_version attribute on the py_binary, not the host
configuration.

Bug: 155429121
Test: NA
Change-Id: I4f73600497cbe5ab89e411471d05c4cbf2e54db6