[BA] Don't show BA link after first CC trial build
During configuration cache trial builds flow we run two builds
in a row and it is likely that both outputs are visible in build
tab. This can lead to the confusion described in the bug.
Build analyzer link is never needed in the first trial build
output so don't show it. It is the easiest improvement on this bug.

Bug: 190572884
Test: manual
Change-Id: I3f80e6ea7fbbe6151603547135ed9267f5583a6c
(cherry picked from commit 911cd87750bd572307fc9d6676eb75bee69a3d32)