CWM-2940: RdConverter for DeviceAndSearch panel (android logcat)
CWM generic toolwindow framework does not support GroupLayout yet. Do manual conversion of the DeviceAndSearchPanel to linear-like layout. This conversion is an approximation: on the guest there is no distinction between wide/narrow cases, and all the widgets (except the checkbox) are always scaled proportionally. (on the host they are scaled proportionally up to 500px, then only search field is scaled)

GitOrigin-RevId: 2e8f9c3e453852bc5b7cb00e13a970f41fa8da81