APK Sign: do not extract passwords in parallel
There is no reason in running two separate postponed password retrieving
operations in parallel, we can retrieve both of them one by one
in single job.
On the contrary running these two operations in parallel is not thread
safe and seem to be the reason of passwords being wrongly decoded by
KeePass. I can not reproduce the issue after this fix and it was easily
reproducible before.

Bug: 192538185
Test: manual
Change-Id: Ie4b691506e3b0a5ab6fa06d7d9428fe208a29aec