Always return null testPackageName for non-test run configurations
`ApplicationIdProvider.testPackageName` is supposed to be used with
test run configurations only.

With this change we always return `null` if the run configuration is not
a test run configuration. The only existing usage that calls this method
on a provider obtained from a non-test run configuration is
`ConnectDebuggerTaskBase` and it in fact expects the result to be null
in such cases.

Bug: n/a
Test: n/a
Change-Id: I6e01463595021cbd26cad5f402686c0384d38b1e