[Dsl] Be more robust to invalid syntax
Finding a match using the INDEX_PATTERN used does not explicitly
distinguish between matching a property and matching a dereference.
This means that some invalid syntaxes could continuously match the
INDEX_PATTERN, leading to attempts to insert null into a collection data
structure.

Add checks against this, and furthermore remove the checks against
groupCount() whose intent was probably to check whether a particular
branch of the match had succeeded but instead check a static property of
the matcher.

This fixes the symptoms shown in EA-329164 from JetBrains.

Bug: none
Test: added
Change-Id: Iae67501dc583719640f09bd6249d02f26b50bc40