[Dsl] Don't add non-expressions to ExpressionLists
In resolving qualified expressions to DslElements, lookup of some
special blocks is hard-coded.  Those lookups should only resolve to
special blocks (or projects) if they are the first
lookup (i.e. implicitly on the project) rather than all the
time.  (TODO: actually they should succeed if the resultElement is a
project).

The treatment of `ext` here is not correct in that it does not support
extra properties on ExtensionAware elements other than the project.
However, in that respect it now mirrors the treatment of extra
properties elsewhere in resolution, so even if the number of sources of
confusion remains the same, there are fewer distinct kinds of confusion
overall.

This fixes the symptom exhibited in EA-297573 from JetBrains.

Bug: none
Test: added
Change-Id: I57bfb53b3851266539729648137007a8405d5f4a