Do not register after-build tasks
and instead listen to the built result futures.

This change is not pure refactoring as with this change multi invocation
builds are handled by invocing the task multiple times instead of for
the first Gradle invocation only. It may result in multiple
notifications which is anyway better than notifying for after and for
the first part of built only.

We use the `directExecutor()` to trasnform futures to handle results
in the same thread they used to be handled before this change.

The newly introduced `for` loops and future transformations will be
further simplified in upcoming changes.

Bug: n/a
Test: n/a (means no tests at all)
Change-Id: I42c3474ca6345450311f6d9ff1ad7ef75c594b66