Re-work BuildsToPathsMapper to consume AssembleInvocationResult directly
This is to hide build system internal details in the project system
implementation module and allow further simplification of the IDE Gradle
models.

Even though with this change, in the caseof a composite build,
`BuildsToPathsMapper` receives the result of the whole build it doesn't
mean that it supports composite builds with AGP 4.0 or older. The
implementation is still unaware of composite builds and uses simple
Gradle project paths. We will try to improve this in later changes.

We also remove outdated unit tests concerned with AGP 2.x support.

Bug: n/a
Test: n/a
Change-Id: I5b30dd0660f6384dcce58aae7222721b59934cb2