[AUA] Inline and remove ..._TOOL_WINDOW flag
While the tool window may or may not be the optimal UI, future
development is likely to use this as a starting point rather than the
previous simple dialog-box UI, which is not rich enough to express the
set of operations that users need to be able to do.  The value of
keeping the flag around is very limited.

Bug: n/a
Test: existing tests adjusted
Change-Id: Id192ee031c341684466e6cdddb03cd804f924adf