[Sync] Do not allow import of an already-open project
Having two open Studio sessions on the same project is likely to lead to
all sorts of issues.  Because the "Import" action unconditionally opens
the project in a new window, the platform-based check for already-open
projects is elided, so explicitly run that check (using the project base
directory rather than the Gradle build file) before calling
ProjectUtil.openOrImport().

Bug: 193486492
Test: manual
Change-Id: I71140ab699f532401e7a496ab197197fa104757c