Fix the project loaded by testExistingGradlePropertiesGroovy
This test was using the KTS project, instead of the Groovy-based one.
Gradle sync using KTS takes at least 5 minutes, so this greatly bloats
test time (groovy sync is ~3 minutes faster).

Now that all tests in MigrateToAndroidxGradleGroovyTest are groovy-based, the
test times should be stable enough to run in the __other split target.
We no longer need to separate this test in its own target.

Bug: 157752747
Test: NA
Change-Id: Iab84ee9b83e71a5c57598574ab5fe6151172bd94