Introduce explict library names in IDE library models
This is to allow library names to come from the sync code rather than be
derived from the artifact addresses when setting up the project. This is
required to support libraries with multiple artifacts for different
variants, capabilities etc., e.g. a test fixtures artifact for "a:b:1.0"
should get a different library name than the main artifact for the same
"a:b:1.0".

Bug: 37797956
Test: Project snapshots
Change-Id: Ic125fa830d7edc5d49b6a91f0f460b22d6a009c3