[DM] Fix deleting AVD in virtual tab
I broke the delete action when I refactored VirtualDisplayList recently,
where I changed setWidths to be called by a TableModelListener. But this
resulted in an IndexOutOfBoundsException when deleting an AVD, because
the listeners are notified in reverse order, which means it was hitting
my custom listener before the listener that updates the table's
RowSorter, so it thought there were more rows than there actually were.
So revert this small piece back to manually calling it from the
VirtualDeviceModelListener.

Bug: 179803112
Test: None
Change-Id: I7af15e83a21b31eaa78471fd95afb183e21b74fd