[kotlin] Use the correct qualified name of the KaptGradleModel (KTIJ-19590)
This is a temporary patch in order to fix Gradle import with the Android plugin enabled.
The proper fix would be depending on the original classes of the Kotlin IDE plugin, or using the newer 'org.jetbrains.kotlin:gradle-tooling' dependency.

KaptGradleModel was moved in 42c8c347c095253115314f31fd251f4dbc66de0c during the Kotlin plugin migration to the new plugin model.

GitOrigin-RevId: f519b4910eedefea97586d39a6b86b355238af39