[Dsl] Support platform and enforcedPlatform dependencies
Treat them as ArtifactDependencies, essentially discarding the
information about them being platform dependencies.  This allows all
downstream code to simply work (e.g. the Project Structure Dialog will
offer a suggestions if the platform version is out-dated; the
Dependencies tab of the PSD will display it and allow changing its
configuration or version) but loses the information about it being a
platform dependency.

Bug: 190076622
Test: added
Change-Id: Ibdcfff6452de1a5e83bc1225780b4f2fae66077a