[Dsl] Define an interface for platform dependencies
We cannot continue to pretend that platform dependencies are always
artifact dependencies, because platforms dependencies can also be
declared on other projects (modules).  Instead, define an interface for
platform dependencies, whose only method in the interface is a boolean
to determine whether the platform dependency is enforced or not.

Subclass the existing ArtifactDependencyModel implementation to be able
to support the new interface; adapt the factory methods on MapNotation
and CompactNotation to return an instance of the subclass where
appropriate.  A subsequent CL will implement the interface for
ModuleDependency.

Bug: 190076622
Test: extended existing test
Change-Id: Id35a388d10b0554042ef109cb1295daa717122e6