Add transform to androidx ResourcesCompat class
The loadFont method assumes that font file paths will always start
with `res/`. That is not the case in Studio. The transform replaces
the call to `startsWith` with a call to `contains` so that the check
passes.

This reverts commit 06f8370623b63c4a5f675cb4f9c070addea26ac0.

Reason for revert: Layoutlib has been updated to fix the issue that
required the original revert.

Bug: 150587499
Change-Id: I31b5587ed252599ba7d9f0c3269cf95388ad02b0
Test: N/A