Optimize Message Processing
Logcat messages arrive from adb logcat in chunks of about 16kb which is about 100 log entries on average.

Under extreme conditions, the backlog of entries in logcat can exceed 1 million. If we send each batch to the EDT as it arrives, the EDT quickly becomes a bottleneck.

The old Logcat code has a 200ms delay on flushing to the UI thread which on my machine results in batches of about 2000 messages per flush.

We use a slightly more complicated strategy:
* We flush if there are no pending messages (receiving channel is empty)
* We flush if we have more than MAX_MESSAGES_PER_BATCH messages accumulated
* We flush if more than MAX_TIME_PER_BATCH_MS milliseconds elapsed since the last time we flushed.

I've experimented with a few values and landed on:
   MAX_TIME_PER_BATCH_MS = 100
   MAX_MESSAGES_PER_BATCH = 5000

These values may change as we learn more and perhaps we can make them server flags, or maybe derive them based on machine specs.

For comparison, with these values, the code renders 466,785 messages in 12 seconds while Logcat V1 loads them in 43 seconds.

Bug: 191880898
Test: Added
Change-Id: I497991ae602ef6a92d72c4919936468db8dadfd5