Improve Device File Explorer menu items text
To reflect multiple nodes being selected. The "Save As..." may have been
especially confusing for users because selecting multiple nodes will
open a dialog that asks the user to open a folder, where the files would
be saved to, rather than the typical "Save As" dialog. Changing the text
to "Save To..." may make it a little easier for the user to understand
that they are being asked for a directory.

Fixes: 150896356
Test: DeviceExplorerControllerTest
Change-Id: I53800b849ed61a2212ec101302fbf28ea995e30f