Rename IntellijNavSource as ApkNavSource
Now that all the other nav sources have been extracted, the only logic
left in IntellijNavSource was the code to search an APK. The class has
been renamed to better reflect the scope of the nav source.

Unlike the other nav sources, this one was left in the original language
(Java) with minimal alterations since we cannot easily test the logic
here.

Bug: 197112399
Test: Untested, no method to easily test APK code navigation.
Change-Id: Ifaa7fd661bd3b7953b63f5a194453289cf3260c9