IDEA-275522 [merge-4.3] move registration of AndroidModuleDataService from p-s-gradle-sync to p-s-gradle
This addon to "Yuriy Solodkyy* 01.04.2021, 15:13 Consolidate (partly) IDE sync code in project-system-gradle" which moved the class (AndroidModuleDataService) but not its registration.

Goal is to remove dependency on intellij.platform.externalSystem, which is not supposed to be used from external build process. For example, externalSystem module is compiled against JAVA11, but build process needs at most 8.

GitOrigin-RevId: 3cae7c9f60e96ad12c073e1cae21c79b19f3d6f4