[Dsl,.toml] Implement convertToPsiElement
Partially, at least.  This is necessary for parsing because
FakeArtifactElements, representing semantically-meaningful portions of
literal strings, have to generate PsiElements corresponding to the value
coming from that portion, in order for the value extraction
functionality to work.

This was not necessary to support a single string as a catalog alias
target, where the PsiElement is generated somewhat dubiously in the
referencing scope -- a build file -- nor for a group/name/version map
as a catalog alias target, where there are no fake elements.  But the
Version Catalog also allows expressing a target artifact as a map
containing a module string (aggregating group and name) and a version,
and in this case the referencing scope is the Toml file, not the build
file.

Bug: 200280395
Test: exercised in next CL
Change-Id: Id6a05c6831821cc84418795425e5ba24beee9aa7