Fix regression in Android coroutines initialization
The kotlin coroutines core library will detect that is running in
Android and try to load "AndroidDispatcherFactory" to create the
coroutines dispatchers. By making it fail, we force the
kotlinx.coroutines.internal.FastServiceLoader to fallback to the backup
loading method. This method involves trying to find the right dispatcher
factory by locating the
kotlinx.coroutines.internal.MainDispatcherFactory service and
instantiating it.
We have our own implementation of that service and we declare it as part
of the layoutlib-extensions module that is loaded as an external jar.
The service will load the LayoutlibDispatcherFactory that implements a
dispatcher adapter to run on layoutlib.

Test: Added regression test
Bug: 162056408
Change-Id: I546474c415c5eb26a2f63e217cf94802a7b3c224