LE: Add test and harness for LE Code Generation
This CL introduce a way to test LE generated JVM bytecode.
It tests AndroidLiveEditCodeGenerator output by invoking the
generated method.

This might not work in the future when we start generating
invalid .class file. We migh have need to keep an option to
generate valid file for testing purpose later.

Test: This add tests
Bug: none
Change-Id: I889eb5b872b2d313c9c5abe2ff4b82a518cd40ad