Reenable testGotoDeclaration, fix ApplicationRule
It turns out that change: ag/16124290 ends up messing up other tests.
The LoadingState is kept in a static that can affect other tests.
When we set the state to "BOOTSTRAP" other services are not loaded.
For testGotoDeclaration the KeymapManager is not initialized and
therefore no keys are loaded for the global GotoDeclaration action.

Revert ag/1612490 to fix testGotoDeclaration, and install a mock for
the PerformanceWatcher that will not cause problems in tests using
ApplicationRule.

Fixes: 204351118
Test: LayoutInspectorTreePanelTest.testGotoDeclaration
Change-Id: Ic06665676bb475d6b634473c8ca1bc06190e9769