[Sync] Detect projects using Version Catalogs
We will need to show the user notices if they are attempting to use
version catalogs, either to confess that Studio does not support them,
or to warn them that any support Studio does have is experimental (at
least to start with).

Detect a Version Catalog project in a way that attempts to be
future-proof given the expected timeline for the feature to be declared
stable in Gradle: the feature needs to be turned on explicitly until the
expected stable version; if either the feature is turned on or the
version of Gradle is higher than the expected stable version, detect use
of the API in the settings file or the implicit use in the filesystem if
no explicit use is detected.

Since this might get called often, at keystroke latency, make sure to
check for cancellation and to cache the result as much as possible;
invalidate cached results only if there are changes to the three files
that are overwhelmingly most likely to matter.

Bug: 200280395
Test: added
Change-Id: If00d2673fbf355749011b6a0e7caa2cd06f9638d